+++ Assets/Scripts/Helpers/MathHelper.cs
@@ -14,10 +14,10 @@ public class MathHelper{
 	
 	}
 
-    public static double degreeBetween2Points(Vector3 p1, Vector3 p2)
+    public static float degreeBetween2Points(Vector3 p1, Vector3 p2)
     {
         float xDiff = p2.x - p1.x;
         float yDiff = p2.y - p1.y;
-        return Math.Atan2(yDiff, xDiff) * 180.0 / Math.PI;
+        return (float)(Math.Atan2(yDiff, xDiff) * 180.0 / Math.PI);
     }
 }
